Dataminning/requirements.txt
appnope==0.1.0
astroid==2.2.5
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
bleach==3.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cycler==0.10.0
dash==1.7.0
dash-core-components==1.6.0
dash-html-components==1.0.2
dash-renderer==1.2.2
dash-table==4.5.1
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.7
django-common-helpers==0.9.2
django-cron==0.5.1
django-crontab==0.7.1
django-mongodb-engine==0.6.0
djangotoolbox==1.8.0
dnspython==1.15.0
entrypoints==0.3
Faker==0.8.13
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Compress==1.4.0
Flask-Cors==3.0.3
Flask-JWT-Extended==3.7.0
Flask-Login==0.4.0
future==0.18.2
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
lightgbm==2.3.1
lxml==4.4.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
missingno==0.4.2
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.4
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
plotly==4.4.1
pluggy==0.6.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
py==1.8.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.7.2
pyparsing==2.4.5
pypi==2.1
pyrsistent==0.15.5
pytest==3.3.0
pytest-flask==0.10.0
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
retrying==1.3.3
scikit-learn==0.22
scipy==1.4.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.2
tornado==6.0.3
tqdm==4.40.2
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
virtualenv==16.7.8
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
